[design]: #7 GNB 아이콘 수정
