--- src/features/main/ui/GNB.tsx
@@ -1,26 +1,0 @@
-export default function GNB() {
-  return (
-    <div className="bg-yellow-300 px-[27px] py-[18px] flex justify-between items-center shadow-[0px -6px 14px 0px rgba(47, 47, 47, 0.04)]">
-      <div className="flex flex-col gap-1 items-center justify-center">
-        <img src="/icon/magazine.svg" alt="매거진" className="w-[26px] h-[26px]" />
-        <p className="text-[#9C9CA1B2] text-[10px]">매거진</p>
-      </div>
-      <div className="flex flex-col gap-1 items-center justify-center">
-        <img src="/icon/magazine.svg" alt="매거진" className="w-[26px] h-[26px]" />
-        <p className="text-[#9C9CA1B2] text-[10px]">적립 내역</p>
-      </div>
-      <div className="flex flex-col gap-1 items-center justify-center">
-        <img src="/icon/magazine.svg" alt="매거진" className="w-[26px] h-[26px]" />
-        <p className="text-[#9C9CA1B2] text-[10px]">홈</p>
-      </div>
-      <div className="flex flex-col gap-1 items-center justify-center">
-        <img src="/icon/magazine.svg" alt="매거진" className="w-[26px] h-[26px]" />
-        <p className="text-[#9C9CA1B2] text-[10px]">스탬프북</p>
-      </div>
-      <div className="flex flex-col gap-1 items-center justify-center">
-        <img src="/icon/magazine.svg" alt="매거진" className="w-[26px] h-[26px]" />
-        <p className="text-[#9C9CA1B2] text-[10px]">마이페이지</p>
-      </div>
-    </div>
-  );
-}
